[Still unoptimized] Read hexadecimal files
This commit adds support for opening hexadecimal ROM files.
Hexadecimal ROM files are files where the contents of the ROM
are not binary, instead they are plain text files where characters
are all hexadecimal, this is, 0123456789ABCDEF.

The emulator will parse the file, convert it to binary data, then
execute it just like it uses to do with binary ROMs.

This is not optimized as it might be buggy.