Introduce abstract keyboard poller
In order to make a library from the CHIP-8 source code, I still
have to decouple the core library from any SDL code. To avoid so,
this commit introduces a generic "keyboard poller".

A poller function has the following form:  char -> int.
Given a char with the number of key the machine wants to know
if pressed or not, the function should return 0 or not 0 depending
on whether the key is pressed or not.

keyboard.h is dropped from cpu.c thanks to this commit.