Clean up main opcode table
The main nibble table is the table that points to 16 functions,
each one covering different opcodes depending on the value for
the most significant nibble, such as the entry #0 for opcodes
whose most significant nibble is 0 or #10 for opcodes whose
most significant nibble is A.

Clean up has been made. Organizing the format and renaming the
functions so that they use a more concise and short format.
Things that have not been done in this commit:

* Fix bugs: no bugfixing here. Period.
* Organize other function blocks: one block per commit.
* Rename datatypes or array types: not yet.
* Replacing many arguments with fewer arguments.