Clean up main loop logic
Not rewritten but many things have changed. Mainly SDL2 is not so
tightly coupled to the main loop now, since most of the logic
that is part of SDL2 has been encapsulated in sdl.c by itself.
Main function only uses init / destroy / render functions but
almost all SDL context is now private to the file.