Safe sound enabling
The application won't crash again if sound cannot be enabled.
Since it is not critical it has been changed to a separate function
and the main method tests if sound can be enabled. If there is
no sound on system or any other kind of error with sound the game
will just be muted.

Closes #19.