Fix DRAW unit test (and endianness, apparently)
So, just like scroll, it was wrong the unit test for DRAW ESM
because wrong variable names. Also I missed a few input parameters
such as I register.

However, I have also changed the actual source code because it
turns out the endianness is not checked and most of the time will
be wrong.

I need a new unit test for checking the endianness of the DRAW
operator. Maybe something that renders a checked bitmap checking
that the correct pixels turn on or off.