Add HDPI support (ESM mode)
Texture has been scaled to 128x64 for making it work with the new
ESM mode, whose screen is 128x64 pixels wide. When ESM is on,
the screen is rendered at actual size (things are smaller but it is
OK).

When ESM mode is off, which is the legacy 64x32 mode, things will be
scaled 200% to make it look fullscreen on a 128x64 display mode.