Conditionally enable tests on presence of check
Running ./configure on a host without libcheck installed won't result in
an error again since libcheck is now an optional dependency. If libcheck
is not found by pkg-config in the host, the tests/ directory won't be
enabled.

Running `make check` on a host that lacks libcheck will result on the
targets (lib8 and chip8) built as usual, but no actual testing will be
made. If the user installs libcheck and re-runs ./configure, the tests
will be compiled as usual.

Thanks to this change, it is now possible to compile the emulator only
relying on libsdl2, which of course must be installed yet.