Dynamic clock cycles
It is now possible to set the number of cycles to run per frame using
the new --speed argument. By default, this will be 16 for compatibility
with what has already been done, but it is possible to set this to a
lower or higher value.

The render update loop has been updated to use a new algorithm that
allows to keep this amount of milliseconds in mind.

(The official number of cycles to run per frame may vary between
emulators, some of them use 7, some of them use 8, some of them use 16
because that is closer to 16.66 to run 1000 cycles per second.)